clean scoring rules code and add helper file with cleanObject function
