+++ src/pages/ScoringRules/RulesFilter/useRulesFilter.tsx
@@ -8,6 +8,7 @@ import type {
 import { getDropDownsValue } from "../../../services/dropdownServices";
 import { useForm } from "react-hook-form";
 import type { RuleIdentifierInterface } from "../rulesServices";
+import { cleanObject } from "../../../utils/helpers";
 
 export interface ViewRulesFormValues {
   id?: number;
@@ -168,30 +169,6 @@ export const useRulesFilter = (
       ]);
     }
   }, [selectedAspect]);
-
-  const cleanObject = <T extends object>(obj: T): Partial<T> => {
-    const newObj: Partial<T> = {};
-    for (const key in obj) {
-      if (Object.prototype.hasOwnProperty.call(obj, key)) {
-        const value = obj[key];
-        if (typeof value === "string" && value !== "") {
-          newObj[key] = value;
-        } else if (typeof value === "number" && value !== 0) {
-          newObj[key] = value;
-        } else if (
-          typeof value === "object" &&
-          value !== null &&
-          !Array.isArray(value)
-        ) {
-          const cleanedSubObject = cleanObject(value as object);
-          if (Object.keys(cleanedSubObject).length > 0) {
-            newObj[key] = cleanedSubObject as T[Extract<keyof T, string>];
-          }
-        }
-      }
-    }
-    return newObj;
-  };
 
   const onSubmit = (data: ViewRulesFormValues) => {
     const filteredData = cleanObject(data);
